+++ packages/flutter_solo/lib/flutter_solo.dart
@@ -8,3 +8,4 @@ export 'package:flutter/foundation.dart' show ValueListenable;
 export 'package:solo/solo.dart';
 export 'src/solo_listenable.dart';
 export 'src/solo_selection.dart';
+export 'src/subscription.dart';